About mobile devices
Hardcoded 6vh, removed detections

Hardcoded the position of the chat input field for mobile preview, regardless of env, so that is visible on mobile devices. Can't find a reliabl way to detect the sizes of address bar and bottom navigation bar.

Removed browser detection from the log, even MDN says it's a very hard problem to solve.